Feasability_Region_Viz1
the objective of the problem was to create the objective function for the desired optimization and define the constraints associated with the problem (y>=2, 2y<=25-x, 4y>=2x-8, and y<=2x-5). The output is a graph displaying the area within the intercepted lines as described by the linear equations (constraints). Identifying the optimal solution using the graphical method suggest the solution is at the interception point of 2 or more of the lines.